+++ lib/pages/admin_profile_page.dart
@@ -8,11 +8,19 @@ import 'package:fitnessapp/widgets/profile_user_stats_graph.dart';
 import 'package:flutter/material.dart';
 import 'package:widgets/widgets.dart';
 
-class AdminProfilePage extends StatelessWidget {
+class AdminProfilePage extends StatefulWidget {
   const AdminProfilePage({super.key});
 
   @override
+  State<AdminProfilePage> createState() => _AdminProfilePageState();
+}
+
+class _AdminProfilePageState extends State<AdminProfilePage>
+    with AutomaticKeepAliveClientMixin {
+  @override
   Widget build(BuildContext context) {
+    super.build(context);
+
     User? currentUser = UserRepository.currentUser;
     return ListView(
       padding: const EdgeInsets.all(20),
@@ -87,4 +95,7 @@ class AdminProfilePage extends StatelessWidget {
       ],
     );
   }
+
+  @override
+  bool get wantKeepAlive => true;
 }