+++ lib/widgets/user_profile_friends_widget.dart
@@ -1,5 +1,5 @@
 import 'package:fitnessapp/database/database.dart';
-import 'package:fitnessapp/models/src/friend.dart';
+import 'package:fitnessapp/models/models.dart';
 import 'package:fitnessapp/pages/home/home_screen.dart';
 import 'package:fitnessapp/widgets/profile_user_stats_graph.dart';
 import 'package:flutter/material.dart';
@@ -7,8 +7,10 @@ import 'package:widgets/widgets.dart';
 
 class UserProfileFriendsGraphPopup extends StatelessWidget {
   final Friend friend;
+  final Future<List<WorkoutStatistic>> loader;
   const UserProfileFriendsGraphPopup({
     required this.friend,
+    required this.loader,
     super.key,
   });
 
@@ -18,25 +20,27 @@ class UserProfileFriendsGraphPopup extends StatelessWidget {
       mainAxisSize: MainAxisSize.min,
       children: [
         ProfileUserStatsGraph(
-          interpretation: '${friend.displayName}\'s workout statistics',
-          loader: UserRepository.getWorkoutDatesStatistics(
-            friend.uid,
-          ),
+          //interpretation: '${friend.displayName}\'s workout statistics',
+          loader: loader,
         ),
-        Align(
-          alignment: AlignmentDirectional.centerEnd,
-          child: TextButtonWidget(
-            'Remove Friend',
-            onPressed: () async {
-              await UserRepository.removeFriend(friend);
-              Navigation.flush(
-                widget: const HomeScreen(
-                  initialIndex: 3,
-                ),
-              );
-            },
-            foregroundColor: context.config.errorColor,
-          ),
+        Row(
+          crossAxisAlignment: CrossAxisAlignment.center,
+          children: [
+            TextWidget(friend.displayName),
+            const Spacer(),
+            TextButtonWidget(
+              'Remove Friend',
+              onPressed: () async {
+                await UserRepository.removeFriend(friend);
+                Navigation.flush(
+                  widget: const HomeScreen(
+                    initialIndex: 3,
+                  ),
+                );
+              },
+              foregroundColor: context.config.errorColor,
+            ),
+          ],
         ),
       ],
     );